Remove TS3.6 support from babel-types
TS3.6 is no longer supported on DT, so babel-types doesn't need to support it anymore either.
